requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.0.1
Django==4.1.5
django-environ==0.9.0
djangorestframework==3.14.0
environ==1.0
filelock==3.9.0
huggingface-hub==0.11.1
idna==3.4
joblib==1.2.0
mysql==0.0.3
mysqlclient==2.1.1
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.0
pandas==1.5.2
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
scikit-learn==1.2.0
scipy==1.10.0
six==1.16.0
sklearn==0.0.post1
sqlparse==0.4.3
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.1
tqdm==4.64.1
transformers==4.25.1
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
